Rectify imports after move
